New deploy script now correctly renames directories and certain references to DarkestHourDev after copying files to content directory.

+++ tools/steamworks/deploy.py
@@ -3,6 +3,7 @@ import os
 import fnmatch
 import re
 import shutil
+import sys
 
 argparser = argparse.ArgumentParser()
 argparser.add_argument('dir', default='.', help='root directory')
@@ -13,15 +14,27 @@ args.dir = os.path.abspath(args.dir)
 
 content_path = os.path.join(os.path.dirname(__file__), 'sdk', 'tools', 'ContentBuilder', 'content')
 
+if not os.path.exists(content_path):
+    print 'unable to find content directory (%s)' % content_path
+    sys.exit(1)
+
+
 # delete everything in the content path
 for fname in os.listdir(content_path):
-    fpath = os.ptah.join(content_path, fname)
+    fpath = os.path.join(content_path, fname)
     try:
         shutil.rmtree(fpath)
     except OSError:
         os.remove(fpath)
 
-with open(os.path.join(args.dir, args.mod, '.steaminclude'), 'rb') as f:
+steaminclude_path = os.path.join(args.dir, args.mod, '.steaminclude')
+
+if not os.path.exists(steaminclude_path):
+    print 'unable to find .steaminclude file'
+    sys.exit(1)
+
+# read include and ignore patterns
+with open(steaminclude_path, 'rb') as f:
     include_patterns = []
     ignore_patterns = []
     for line in filter(lambda x: len(x) > 0, map(lambda x: x.strip(), f.read().splitlines())):
@@ -30,23 +43,39 @@ with open(os.path.join(args.dir, args.mod, '.steaminclude'), 'rb') as f:
             ignore_patterns.append(m.groups()[0])
         else:
             include_patterns.append(line)
-    for root, dirs, filenames in os.walk(args.dir):
-        for filename in filenames:
-            is_included = False
-            relpath = os.path.relpath(os.path.abspath(os.path.join(root, filename)), args.dir)
-            for include_pattern in include_patterns:
-                if fnmatch.fnmatch(relpath, include_pattern):
-                    is_included = True
-                    break
-            if not is_included:
-                continue
-            for ignore_pattern in ignore_patterns:
-                if fnmatch.fnmatch(relpath, ignore_pattern):
-                    is_included = False
-            if not is_included:
-                continue
-            # make sure subdiretories exist
-            if not os.path.exists(os.path.join(content_path, os.path.dirname(relpath))):
-                os.makedirs(os.path.join(content_path, os.path.dirname(relpath)))
-            # copy file to 
-            shutil.copy(os.path.join(args.dir, relpath), os.path.join(content_path, relpath))
+
+# now walk directory and move files to the content directory
+for root, dirs, filenames in os.walk(args.dir):
+    for filename in filenames:
+        is_included = False
+        relpath = os.path.relpath(os.path.abspath(os.path.join(root, filename)), args.dir)
+        for include_pattern in include_patterns:
+            if fnmatch.fnmatch(relpath, include_pattern):
+                is_included = True
+                break
+        if not is_included:
+            continue
+        for ignore_pattern in ignore_patterns:
+            if fnmatch.fnmatch(relpath, ignore_pattern):
+                is_included = False
+        if not is_included:
+            continue
+        # make sure subdiretories exist
+        if not os.path.exists(os.path.join(content_path, os.path.dirname(relpath))):
+            os.makedirs(os.path.join(content_path, os.path.dirname(relpath)))
+        # copy file to 
+        shutil.copy(os.path.join(args.dir, relpath), os.path.join(content_path, relpath))
+
+# rename darkesthourdev to darkesthour in relevant places
+# rename folders and content (OfficialMod, Default.ini)
+for root, dirs, filenames in os.walk(content_path):
+    for dir in filter(lambda x: x == 'DarkestHourDev', dirs):
+        os.rename(os.path.join(root, dir), os.path.join(root, 'DarkestHour'))
+
+for root, dirs, filenames in os.walk(content_path):
+    for filename in filter(lambda x: os.path.splitext(x)[1] == '.ini', filenames):
+        with open(os.path.join(root, filename), 'r+') as f:
+            c = f.read().replace('DarkestHourDev', 'DarkestHour')
+            f.seek(0)
+            f.truncate()
+            f.write(c)
